+++ packages/auth-utils/src/index.ts
@@ -1,4 +1,4 @@
 export { CogniteAuth } from './authentication';
-export { getFlow, saveFlow, removeFlow } from './utils';
+export { getFlow, saveFlow, removeFlow, decodeTokenUid } from './utils';
 export type { AuthFlow, AuthenticatedUser } from './types';
 export * from './fakeIdP';