+++ packages/react-comments/src/utils/getUserIdFromToken.tsx
@@ -1,0 +1,10 @@
+import { decodeTokenUid } from '@cognite/auth-utils';
+
+export const getUserIdFromToken = (idToken?: string): string => {
+  if (idToken) {
+    const userId = decodeTokenUid(idToken);
+    return userId || '';
+  }
+
+  return '';
+};